Method lad workd but needs to be run twice
When clicking on the load button, nothing will happen.

But if it is clicked a second time, the method will work.

This is why I have added fileLoad attribute meanwhile, -> so that if the load button has already been clicked, the user does not need to re-select the file he wants to extract information from.